% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testFunction.R
\name{MyNewFunction2}
\alias{MyNewFunction2}
\title{MyNewFunction
This is a function}
\usage{
MyNewFunction2(argument1)
}
\arguments{
\item{argument1}{This is my parameter}
}
\description{
MyNewFunction
This is a function
}
